requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
pydantic==2.6.1
psycopg2-binary==2.9.9
requests==2.31.0
aiofiles==23.2.1
